pythoncode/DS P1/requirements.txt
Flask==3.0.0
google-api-core==2.12.0
google-api-python-client==2.102.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
googleapis-common-protos==1.60.0
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.4
Jinja2==3.1.2
urllib3==2.0.6